Task management

task_status() retrieves the status of a task started by copy() or transfer().

task_cancel() cancels a task started with copy() or transfer().

task_list() retrieves information on all tasks submitted by the user.

Usage

task_status(.data, all_fields = FALSE)

task_cancel(.data)

task_list(filter = "", orderby = "request_time DESC", all_fields = FALSE)

Arguments

.data

a tibble with column task_id, as returned by copy() or transfer().

all_fields

logical(1) indicating whether abbreviated or detailed information about the task status should be returned

filter

character(1) query parameter restricting which tasks are returned by task_list(). See details.

orderby

character(1) task list ordering; the default places the most recently requested task at the top.

Value

task_status() and task_list() return a tibble summarizing the status of the task(s). The meaning of each column is described in the Globus documentation at

https://docs.globus.org/api/transfer/task/#task_document

With the default value all_fields = FALSE, the tibble contains columns

task_id: the id of the task.

status: one of ACTIVE, INACTIVE, SUCCEEDED or FAILED.

type: one of TRANSFER or DELETE.

nice_status: NULL for a completed task; OK or Queued tasks proceeding normally; or an indication of the reason for ACTIVE or INACTIVE failure. Examples of nice_status indicating transient failure include: PERMISSION_DENIED if the user does not currently have access to the source or destination resource; CONNECT_FAILED if the Globus Connect Personal client is not currently on-line; GC_PAUSED if the client is paused.

task_cancel() returns its input argument, invisibly.

Details

The input to task_status() is usually the return value from copy() or transfer().

task_cancel() prints a message indicating successful submission of the task cancellation operation.

For task_list(), the filter and orderby syntax is described at

https://docs.globus.org/api/transfer/task/#filter_and_order_by_options

A simple task list filter is to return all 'ACTIVE' tasks, with filter = "status:ACTIVE".

Examples

if (exists("task")) {
    ## check transfer task status
    task_status <- task_status(task)
    print(task_status)

    ## tasks that are failing need to be cancelled
    status <- pull(task_status, "status")
    nice_status <- pull(task_status, "nice_status")
    cancel <-
        ## task has failed, or...
        identical(status, "FAILED") ||
        ## ...something's gone wrong
        (!is.null(nice_status) && !nice_status %in% c("OK", "Queued"))
    if (cancel)
        task_cancel(task)
}

task_list()
#> # A tibble: 29 × 5
#>    task_id                              type     status    nice_status label    
#>    <chr>                                <chr>    <chr>     <lgl>       <chr>    
#>  1 b7e8fdfe-5995-11ef-b6d4-f55c894bd1c6 TRANSFER SUCCEEDED NA          2024-08-…
#>  2 7776b014-598f-11ef-b6d4-f55c894bd1c6 TRANSFER SUCCEEDED NA          2024-08-…
#>  3 550bfa46-5987-11ef-b6d4-f55c894bd1c6 TRANSFER SUCCEEDED NA          2024-08-…
#>  4 d2badc10-5986-11ef-b917-c30b78766494 TRANSFER SUCCEEDED NA          2024-08-…
#>  5 ff043bce-5984-11ef-abcf-8b7014cf3b82 TRANSFER FAILED    NA          2024-08-…
#>  6 e337d7de-5984-11ef-abcf-8b7014cf3b82 TRANSFER SUCCEEDED NA          2024-08-…
#>  7 8473e936-5984-11ef-b6d4-f55c894bd1c6 TRANSFER SUCCEEDED NA          2024-08-…
#>  8 4642c662-597c-11ef-b6d4-f55c894bd1c6 TRANSFER FAILED    NA          2024-08-…
#>  9 4a1a86a4-597b-11ef-b6d4-f55c894bd1c6 TRANSFER FAILED    NA          2024-08-…
#> 10 76cc6704-58f8-11ef-b917-c30b78766494 TRANSFER SUCCEEDED NA          2024-08-…
#> # ℹ 19 more rows